LayoutTests/platform/glib/imported/w3c/web-platform-tests/cors/client-hint-request-headers-2.tentative-expected.txt

CONSOLE MESSAGE: Request header field save-data is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors/resources/cors-makeheader.py?headers=x-print, due to access control checks.
Request headers


FAIL Client hint headers are simple headers  A network error occurred.
